integrated to MS VS 2012 and fixed child illnesses

+++ EnergyDeposition.cpp
@@ -2,11 +2,11 @@
 
 
 
-float EnergyDeposition::getEnergyDeposition ( float delta_position )
+double EnergyDeposition::getEnergyDeposition ( double delta_position )
 {
     // muss noch bearbeitet werden. wie genau aussieht.
-            float ENERGYLOSS = 5;
-            float DISTANCE = 1;
+            double ENERGYLOSS = 5;
+            double DISTANCE = 1;
             return (ENERGYLOSS*delta_position/DISTANCE);
 }
